Fix Assistant Details delete CTA
